Relocate board methods to board folder
