+++ lib/focalboard/__tests__/setDatabaseProposalProperties.spec.ts
