+++ lib/focalboard/setDatabaseProposalProperties.ts
