+++ pages/api/pages/[id]/proposal-source.ts
@@ -3,9 +3,9 @@ import { prisma } from '@charmverse/core/prisma-client';
 import type { NextApiRequest, NextApiResponse } from 'next';
 import nc from 'next-connect';
 
+import { createCardsFromProposals } from 'lib/focalboard/createCardsFromProposals';
+import { updateCardsFromProposals } from 'lib/focalboard/updateCardsFromProposals';
 import { ActionNotPermittedError, NotFoundError, onError, onNoMatch, requireKeys, requireUser } from 'lib/middleware';
-import { createCardsFromProposals } from 'lib/pages/createCardsFromProposals';
-import { updateCardsFromProposals } from 'lib/pages/updateCardsFromProposals';
 import { computePagePermissions } from 'lib/permissions/pages/computePagePermissions';
 import { withSessionRoute } from 'lib/session/withSession';
